Add new timed task type
No task that represents a task with a duration but no start or end
date.

Such tasks are very common and should have proper representation.

Creating a subtype of tasks allows us to represent more useful
information for the user.

Let's use a /takes argument to accept the duration in hours.